+++ tox.ini
@@ -28,6 +28,6 @@ deps =
 
 [testenv:lint]
 commands =
-    pylint pyls
-    pycodestyle pyls
-    pyflakes pyls
+    pylint pyls test
+    pycodestyle pyls test --exclude=test/plugins/.ropeproject,test/.ropeproject
+    pyflakes pyls test
